add: サービスクラスに診断スコアロジックを移動
